Function coprocessor_plugin_api::pkgname_to_libname

pub fn pkgname_to_libname(pkgname: &str) -> String

Transforms the name of a package into the name of the compiled library.

The result of the function can be used to correctly locate build artifacts of dylib on different platforms.

The name of the dylib is

lib<pkgname>.so on Linux

lib<pkgname>.dylib on MaxOS

lib<pkgname>.dll on Windows

See also https://doc.rust-lang.org/reference/linkage.html

Note: Depending on artifacts of other crates will be easier with this RFC.